requirements.txt
flask>=3.0
supabase==2.9.1
openai>=1.0
python-dotenv>=1.0
gunicorn>=21.0
tzdata>=2024.1
